+++ compiler/gen_wasm/src/lib.rs
@@ -7,8 +7,6 @@ pub mod opcodes;
 pub mod serialize;
 mod storage;
 
-mod tests;
-
 use bumpalo::{self, collections::Vec, Bump};
 use parity_wasm::builder;
