+++ compiler/test_wasm/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "test_wasm"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+# roc_module = { path = "../module" }
+# roc_mono = { path = "../mono" }
+
+# # TODO: switch to parity-wasm 0.44 once it's out (allows bumpalo vectors in some places)
+parity-wasm = { git = "https://github.com/brian-carroll/parity-wasm", branch = "master" }
+
+wasmer = "2.0.0"
+wasmer-wasi = "2.0.0"
+
+roc_collections = { path = "../collections" }
+roc_std = { path = "../../roc_std" }
+bumpalo = { version = "3.6.1", features = ["collections"] }
+roc_gen_wasm = { path = "../gen_wasm" }
+roc_can = { path = "../can" }
+roc_builtins = { path = "../builtins" }
+roc_load = { path = "../load" }
+roc_types = { path = "../types" }
+roc_module = { path = "../module" }
+indoc = "0.3.3"
+pretty_assertions = "0.5.1"
+libc = "0.2"
+target-lexicon = "0.12.2"
+tempfile = "3.1.0"
