+++ compiler/test_wasm/src/lib.rs
